requirements.txt
tweepy==4.14.0
Pillow==10.0.1
schedule==1.2.0
python-decouple==3.8
pytz==2023.3